requirements.txt
asgiref==3.5.0
boto3==1.21.8
botocore==1.24.8
Django==4.0.2
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
jmespath==0.10.0
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
